fix(web): Show warning for transient-only network setup
Replaced the previous AllConnectionsStatusAlert component with a simpler
NoPersistentConnectionsAlert that displays a warning only when all
network connections are set to be transient (not persisted in the
installed system).

Removed unused complexity such as mixed-mode detection and global
actions, which may be revisited in future iterations.
